+++ tiled/catalog/adapter.py
@@ -347,6 +347,7 @@ class CatalogNodeAdapter:
     async def startup(self):
         if (self.context.engine.dialect.name == "sqlite") and (
             self.context.engine.url.database == ":memory:"
+            or self.context.engine.url.query.get("mode", None) == "memory"
         ):
             # Special-case for in-memory SQLite: Because it is transient we can
             # skip over anything related to migrations.
@@ -1505,14 +1506,19 @@ CatalogNodeAdapter.register_query(Like, like)
 
 def in_memory(
     *,
+    named_memory=None,
     metadata=None,
     specs=None,
     writable_storage=None,
     readable_storage=None,
     echo=DEFAULT_ECHO,
     adapters_by_mimetype=None,
+    top_level_access_blob=None,
 ):
-    uri = "sqlite:///:memory:"
+    if not named_memory:
+        uri = "sqlite:///:memory:"
+    else:
+        uri = f"sqlite:///file:{named_memory}?mode=memory&cache=shared&uri=true"
     return from_uri(
         uri=uri,
         metadata=metadata,
@@ -1522,6 +1528,7 @@ def in_memory(
         init_if_not_exists=True,
         echo=echo,
         adapters_by_mimetype=adapters_by_mimetype,
+        top_level_access_blob=top_level_access_blob,
     )
 
 
@@ -1559,8 +1566,10 @@ def from_uri(
         logger.info(f"Subprocess stderr: {stderr}")
 
     parsed_url = make_url(uri)
-    if (parsed_url.get_dialect().name == "sqlite") and (
-        parsed_url.database != ":memory:"
+    if (
+        (parsed_url.get_dialect().name == "sqlite")
+        and (parsed_url.database != ":memory:")
+        and (parsed_url.query.get("mode", None) != "memory")
     ):
         # For file-backed SQLite databases, connection pooling offers a
         # significant performance boost. For SQLite databases that exist